Update MergeSort.py (#927)
Remove the self parameter from both the merge() and mergeSort() functions as they are not instance methods of the class.
Rename the arr parameter in the mergeSort() function to unsorted_list or something similar to make it more descriptive.
Add type hints to the function parameters and return values to improve code readability and maintainability.
Update the docstring of the mergeSort() function to adhere to the Python docstring conventions.                                                                                                                                                                                                                                     Note: I assumed that the input list contains only integers. If the list can contain any type of elements, you may need to modify the code accordingly.

Signed-off-by: NIKITA PANDEY <113332472+nikitapandeyy@users.noreply.github.com>